[level 1] Title: x만큼 간격이 있는 n개의 숫자, Time: 0.04 ms, Memory: 77.7 MB -BaekjoonHub

+++ 프로그래머스/lv1/12954. x만큼 간격이 있는 n개의 숫자/x만큼 간격이 있는 n개의 숫자.java
@@ -1,8 +1,8 @@
 class Solution {
     public long[] solution(int x, int n) {
         long[] answer = new long[n];
-        for(long i=1;i<=n;i++){
-            answer[(int)i-1]=i*x;
+        for(int i=1;i<=n;i++){
+            answer[i-1]=(long)i*x;
         }
         return answer;
     }
